(상혁) Inquiry DTO, 상혁 서비스, 컨트롤러, 상품상세 board.mapping.xml 수정 업로드

+++ workspace/SlimBear/src/main/java/com/kgitbank/slimbear/dto/InquiryDTO.java
@@ -17,5 +17,5 @@ public class InquiryDTO {
 	private Date reg_date;
 	private String status;
 	private String type;
-	private Integer prod_uid;
+	private Long prod_uid;
 }
